refactor: FSD Phase 6c — 認証・アカウント系の単一画面フックを pages の model へ移す (FRESTYLE-168)
Phase 6b と同じ方針で、認証・オンボーディング・アカウント系の単一画面専用
フック 11 個を pages/<slice>/model へ移す。すべて store・他 hook 非依存。

- useLoginPage → pages/login/model
- useLoginCallback → pages/login-callback/model
- useForgotPassword → pages/forgot-password/model
- useConfirmForgotPassword → pages/confirm-forgot-password/model
- useAcceptInvitation → pages/accept-invitation/model
- useCompanyApplication → pages/company-application/model
- useProfileEdit / useProfileImageUpload / useProfileStats → pages/settings/model
- useUserDashboard / useCompanyLearningSummary → pages/home/model

invitationToken(招待トークンの localStorage 保存)は accept-invitation /
login-callback / entities/user/api/authRepository の 3 箇所で使う共有
ユーティリティ(依存なしの純粋関数)なので shared/lib へ移した。これで
src/lib ディレクトリは空になり削除。

pages 間 import が無いことを確認済み(境界 lint 緑)。既存テストの assert は
変更していない。テスト件数 1184 で不変。
