+++ frontend/src/pages/home/model/__tests__/useCompanyLearningSummary.test.ts
