+++ frontend/src/pages/home/model/useCompanyLearningSummary.ts
