+++ frontend/src/pages/login-callback/ui/LoginCallback.tsx
@@ -1,5 +1,5 @@
 import Loading from '@/shared/ui/Loading';
-import { useLoginCallback } from '@/hooks/useLoginCallback';
+import { useLoginCallback } from '../model/useLoginCallback';
 
 export default function LoginCallback() {
   useLoginCallback();
