+++ frontend/src/pages/settings/model/__tests__/useProfileStats.test.ts
