+++ frontend/src/shared/lib/__tests__/invitationToken.test.ts
