+++ frontend/src/shared/lib/invitationToken.ts
